# libs/contracts/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: DocVault API Gateway
  description: |
    Unified API gateway for DocVault — secure document management system.

    ## Authentication

    **Development (localhost):** After signing in via `/auth/login`, the gateway sets an
    `dv_access_token` cookie. Subsequent API calls automatically include this cookie.

    **Production:** Pass the Keycloak JWT in the `Authorization` header:
    `Authorization: Bearer <token>`

    Both methods are accepted on all protected endpoints.

    ## Roles
    | Role | Permissions |
    |------|-------------|
    | `viewer` | Read published documents, download files |
    | `editor` | Create, upload, submit, archive documents |
    | `approver` | Approve/reject pending documents |
    | `compliance_officer` | Read audit logs and allowed metadata; cannot preview, stream, presign, or download file content |
    | `admin` | Full access |

    ## Document Lifecycle
    ```
    DRAFT → PENDING → PUBLISHED → ARCHIVED
    ```

    ## Compliance Officer Restriction
    Compliance officers (`compliance_officer`) are **never allowed** to preview,
    stream, presign, or download file content, regardless of ACL configuration —
    enforced by route guards and policy.service.ts.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.docvault.internal
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Keycloak-issued RS256 JWT token passed as `Authorization: Bearer <token>`"
    cookieAuth:
      type: apiKey
      in: cookie
      name: dv_access_token
      description: JWT issued by the gateway after Keycloak OIDC login — set automatically by the frontend
  schemas:
...
    ClassificationLevel:
      type: string
      enum: [PUBLIC, INTERNAL, CONFIDENTIAL, SECRET]
      description: |
        Document security classification level.
        - PUBLIC: Anyone authenticated can view and download
        - INTERNAL: viewer+ role, must be PUBLISHED
        - CONFIDENTIAL: editor+ role, PUBLISHED, owner or explicit ACL required
        - SECRET: approver+ role, PUBLISHED, owner or explicit ACL required
        CONFIDENTIAL and SECRET documents are watermarked on download.

    # ─── Documents ─────────────────────────────────────────────────────────────
...
    DocumentVersion:
      type: object
      properties:
        id: { type: string }
        docId: { type: string }
        version: { type: integer }
        objectKey: { type: string }
        checksum: { type: string, description: SHA-256 hex }
        size: { type: integer }
        filename: { type: string }
        contentType: { type: string }
        dlpStatus: { type: string, enum: [NOT_SCANNED, CLEAR, DETECTED] }
        dlpFindings: { type: array, nullable: true, items: { type: object } }
        createdAt: { type: string, format: date-time }
    CreateVersion:
      type: object
      required: [version, objectKey, checksum, filename]
      properties:
        version: { type: integer }
        objectKey: { type: string }
        checksum: { type: string }
        size: { type: integer }
        filename: { type: string }
        contentType: { type: string }
        dlpStatus: { type: string, enum: [NOT_SCANNED, CLEAR, DETECTED] }
        dlpFindings: { type: array, items: { type: object } }
        dlpSuggestedClassification: { $ref: '#/components/schemas/ClassificationLevel' }

    # ─── ACL ──────────────────────────────────────────────────────────────────
...
    PresignedDownloadUrl:
      type: object
      properties:
        docId: { type: string }
        version: { type: integer }
        filename: { type: string }
        expiresInSeconds: { type: integer }
        expiresAt: { type: string, format: date-time }
        url: { type: string, format: uri, nullable: true, description: MinIO presigned download URL. Null when watermarkRequired is true (use streaming endpoint instead). }
        watermarkRequired: { type: boolean, default: false }
        streamingEndpoint: { type: string, nullable: true, description: Path to the streaming endpoint for watermarked downloads. Present when url is null. }
...
paths:
...
  /metadata/documents/{docId}/versions:
    post:
      summary: Register a new document version
      operationId: createVersion
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/docId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersion'
      responses:
        '201':
          description: Version registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersion'
        '403':
          description: Forbidden — editor or admin
        '404':
          description: Document not found
...
  /documents/{docId}/upload:
    post:
      summary: Upload a document blob and register a version
      description: |
        Scans uploads before storage. EICAR malware test content is rejected
        before MinIO/object creation. DLP detections are persisted on the
        version and parent document; PUBLIC/INTERNAL documents with DLP hits
        are escalated to CONFIDENTIAL.
      operationId: uploadDocument
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/docId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Binary file, max 20 MB
      responses:
        '201':
          description: Upload successful, version registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersion'
        '400':
          description: Bad request — file required, exceeds size limit, or malware detected
        '403':
          description: Forbidden — owner editor or admin
  /documents/{docId}/presign-download:
    post:
      summary: Request a presigned download URL
      description: |
        Issues a time-limited presigned URL after download authorization.
        Compliance officers are denied because presign exposes file content access.
      operationId: presignDownload
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Downloads]
      parameters:
        - $ref: '#/components/parameters/docId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: integer }
      responses:
        '200':
          description: Presigned URL issued (valid for 5 minutes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedDownloadUrl'
        '403':
          description: Forbidden — compliance_officer, invalid grant token, expired grant token, or policy denial
  /documents/{docId}/versions/{version}/stream:
    get:
      summary: Stream a document version directly
      description: |
        Returns the document binary as a streaming response.
        If `token` is provided, it must be a valid download grant token bound to this
        document and version. Without `token`, document-service re-authorizes through
        metadata-service before streaming.
        Compliance officers are denied because streaming exposes file content access.

        **Watermarking:** For documents classified as CONFIDENTIAL or SECRET,
        a watermark containing user identity, timestamp, and classification level
        is prepended to the file bytes before streaming.
      operationId: streamVersion
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Downloads]
      parameters:
        - $ref: '#/components/parameters/docId'
        - name: version
          in: path
          required: true
          schema:
            type: integer
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: Optional download grant token issued by `/metadata/documents/{docId}/download-authorize`
      responses:
        '200':
          description: Binary file stream (watermarked for CONFIDENTIAL/SECRET)
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden — compliance_officer, invalid grant token, expired grant token, or policy denial
        '404':
          description: Document or version not found
  /documents/{docId}/preview:
    get:
      summary: Stream a document for inline preview (supports HTTP Range requests)
      description: |
        Streams the raw document binary for inline display in a browser.
        No watermark is applied. Supports HTTP Range requests for PDF.js lazy loading.
        Compliance officers are denied because preview is file content access.
      operationId: previewDocument
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/docId'
        - name: version
          in: query
          required: false
          schema:
            type: integer
          description: Specific version to preview (defaults to current version)
      responses:
        '200':
          description: Binary file stream (inline, no watermark)
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (Range request)
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden — compliance_officer or metadata preview policy denial
        '404':
          description: Document or version not found

  # ─── Workflow ──────────────────────────────────────────────────────────────
  /workflow/{docId}/submit:
    post:
      summary: Submit a document (DRAFT → PENDING)
      operationId: submitDocument
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Workflow]
      parameters:
        - $ref: '#/components/parameters/docId'
      responses:
        '200':
          description: Submitted successfully
        '403':
          description: Forbidden — owner editor or admin
        '409':
          description: Conflict — only DRAFT documents can be submitted
...